Add TileDataEditor
