+++ Assets/Editor/TileDataEditor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a006a4932e8654a0ab8782317f9abca5
+timeCreated: 1500995876
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
