+++ Assets/Mapzen/FeatureStyle.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0eb0cbbdcd42b4a18805177b7dc9da97
+timeCreated: 1500668439
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
